breaking: rename parity module to scatter
all functions were more conveniently importable from __init__.py (e.g. from pymatviz import density_scatter) so hopefully doesn't affect many users
