+++ site/src/routes/[slug]/parity/1.density-scatter.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from pymatviz.io import save_and_compress_svg
-from pymatviz.parity import density_scatter
+from pymatviz.scatter import density_scatter
 
 
 rand_regression_size = 500
