+++ site/src/routes/[slug]/parity/5.scatter-with-err-bar.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from pymatviz.io import save_and_compress_svg
-from pymatviz.parity import scatter_with_err_bar
+from pymatviz.scatter import scatter_with_err_bar
 
 
 rand_regression_size = 500
