+++ site/src/routes/[slug]/parity/_3.density-scatter-hex.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from pymatviz.io import save_and_compress_svg
-from pymatviz.parity import density_hexbin
+from pymatviz.scatter import density_hexbin
 
 
 rand_regression_size = 500
