+++ site/src/routes/[slug]/parity/_4.density-scatter-hex-with-hist.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from pymatviz.io import save_and_compress_svg
-from pymatviz.parity import density_hexbin_with_hist
+from pymatviz.scatter import density_hexbin_with_hist
 
 
 rand_regression_size = 500
